Update of Parkings fields to DecimalFields

+++ ipa/models.py
@@ -4,8 +4,8 @@ from django.db.models.deletion import CASCADE
 
 class Parkings(models.Model):
     location = models.CharField(max_length=100)
-    longitude = models.IntegerField(null=True)
-    latitude = models.IntegerField(null=True)
+    longitude = models.DecimalField(max_digits=10, decimal_places=6, null=True)
+    latitude = models.DecimalField(max_digits=10, decimal_places=6, null=True)
     
     def __str__(self):
         return self.location
